# contracts/openapi/bsmart-v1.yaml
openapi: 3.1.0
info:
  title: bSmart Client API
  version: 1.8.0
  description: Versioned research and legacy account contract. Current native Google/Apple login uses Supabase Auth directly; wallet registry is specified in supabase-wallet.yaml and docs/contracts/supabase_account.md.
servers:
  - url: https://api.bsmart.today
security:
  - installationBearer: []
paths:
...
  /v1/installations:
    post:
      operationId: createAnonymousInstallation
      summary: Create or refresh the anonymous installation session used before account login exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallationRegistration"
      responses:
        "200":
          description: Anonymous installation session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationSession"
        "201":
          description: Newly created anonymous installation session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstallationSession"
  /v1/devices:
    put:
      operationId: putCurrentDevice
      summary: Register or update the current installation's APNs device token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistrationInput"
      responses:
        "204":
          description: Device registration updated
  /v1/notification-preferences:
    get:
      operationId: getNotificationPreferences
      summary: Return alert delivery preferences for the current installation.
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      operationId: putNotificationPreferences
      summary: Replace alert delivery preferences for the current installation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
  /v1/daily-digest:
    get:
      operationId: getDailyDigest
      summary: Return the latest immutable daily portfolio brief for the current installation.
      responses:
        "200":
          description: Latest generated daily digest snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyDigestSnapshot"
        "404":
          description: No daily digest has been generated for this installation yet
  /v1/mr-collie/query:
    post:
      operationId: queryMrCollie
      summary: Ask Mr Collie to explain the current portfolio and ticker evidence using DeepSeek.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MrCollieQuery"
      responses:
        "200":
          description: Evidence-grounded answer with server-validated citations.
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always `no-store` because the response is installation-specific.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrCollieResponse"
        "429":
          description: Installation request limit reached.
        "502":
          description: DeepSeek returned an unusable response.
        "503":
          description: DeepSeek is not configured or temporarily unavailable.
  /v1/telemetry/events:
    post:
      operationId: postClientTelemetryEvents
      summary: Record a privacy-minimized batch of anonymous product interaction events.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientTelemetryBatch"
      responses:
        "202":
          description: Telemetry batch accepted idempotently
...
  /v1/portfolio/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: putPortfolioEntry
      summary: Idempotently create or replace a locally identified manual position or watchlist entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioEntryInput"
      responses:
        "200":
          description: Upserted portfolio entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioPosition"
    delete:
      operationId: deletePortfolioEntry
      summary: Delete a manual position or watchlist entry.
      responses:
        "204":
          description: Portfolio entry deleted
...
  /v1/feed:
    get:
      operationId: getPortfolioSignals
      summary: Return decision-ready Smart Account and Smart Money signals for the current portfolio.
      parameters:
        - { name: If-None-Match, in: header, schema: { type: string } }
      responses:
        "304":
          description: Authenticated read model is unchanged; reuse the cached response body.
        "503":
          description: Publication changed during this read; retry shortly.
        "200":
          description: Portfolio signals
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioSignal"
  /v1/signals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getPortfolioSignal
      summary: Return one decision-ready portfolio signal with its evidence snapshot.
      responses:
        "200":
          description: Portfolio signal
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSignal"
  /v1/signals/{id}/state:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: putPortfolioSignalState
      summary: Replace the current user's read, saved, ignored, and feedback state for one signal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalUserStateInput"
      responses:
        "200":
          description: Updated signal state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalUserState"
  /v1/smart-account-updates:
    get:
      operationId: getSmartAccountUpdates
      summary: Return material view updates from ranked offchain investment creators.
      parameters:
        - { name: If-None-Match, in: header, schema: { type: string } }
      responses:
        "304":
          description: Authenticated read model is unchanged; reuse the cached response body.
        "503":
          description: Publication changed during this read; retry shortly.
        "200":
          description: Smart Account updates
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartAccountUpdate"
  /v1/smart-accounts/{accountId}/evidence:
    parameters:
      - name: accountId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getSmartAccountEvidence
      summary: Return bounded historical Call evidence for one formally ranked Smart Account.
      parameters:
        - { name: If-None-Match, in: header, schema: { type: string } }
      responses:
        "304":
          description: Authenticated evidence is unchanged; reuse this URL's cached response body.
        "503":
          description: Publication changed during this read; retry shortly.
        "200":
          description: Recent, representative successful, and representative missed calls.
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartAccountUpdate"
  /v1/smart-money-movements:
    get:
      operationId: getSmartMoneyMovements
      summary: Return material position changes from ranked public onchain accounts.
      parameters:
        - name: ticker
          in: query
          schema: { type: string }
        - name: account_id
          in: query
          schema: { type: string }
        - name: after
          in: query
          description: Return movements strictly newer than this timestamp.
          schema: { type: string, format: date-time }
        - name: before
          in: query
          description: Return movements strictly older than this timestamp for cursor-style history replay.
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 250
      responses:
        "200":
          description: Smart Money movements
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
            X-BSmart-Result-Count:
              schema: { type: integer }
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartMoneyMovement"
  /v1/smart-money/{accountId}/evidence:
    parameters:
      - name: accountId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getSmartMoneyEvidence
      summary: Return up to three representative observed entry markets for one public capital account.
      responses:
        "200":
          description: Representative entry timing with exact-contract Hyperliquid price evidence.
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartMoneyRepresentativeEvidence"
...
  /v1/tickers/{symbol}/research:
    get:
      operationId: getTickerResearch
      deprecated: true
      summary: Return a legacy combined research view for one supported ticker.
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticker research summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerResearch"
...
  /v1/tickers/{symbol}/intelligence:
    get:
      operationId: getTickerIntelligence
      summary: Return the current two-source intelligence snapshot for one supported ticker.
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticker intelligence snapshot
          headers:
            ETag:
              $ref: "#/components/headers/ReadModelETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerIntelligence"
...
components:
...
  securitySchemes:
    accountBearer:
      type: http
      scheme: bearer
      description: Opaque verified account session, distinct from installation authentication.
    installationBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-installation-token
  schemas:
...
    InstallationRegistration:
      type: object
      required: [installationId, platform, appVersion, locale, timeZone]
      properties:
        installationId: { type: string, format: uuid }
        platform:
          type: string
          enum: [ios]
        appVersion: { type: string }
        locale: { type: string }
        timeZone: { type: string }
    InstallationSession:
      type: object
      required: [installationId, accessToken, expiresAt]
      properties:
        installationId: { type: string, format: uuid }
        accessToken: { type: string, minLength: 1 }
        expiresAt: { type: string, format: date-time }
    DeviceRegistrationInput:
      type: object
      required: [apnsToken, environment, appVersion, locale, timeZone]
      properties:
        apnsToken: { type: string, minLength: 1 }
        environment:
          type: string
          enum: [development, production]
        appVersion: { type: string }
        locale: { type: string }
        timeZone: { type: string }
...